refactor(worker): extract agent-turn claim phase
Move claim, policy, admission, and turn.started wiring into claimTurnAttempt.
The orchestrator continues from the claimed turn bag.
